+++ src/main/java/baseball/version1/Computer.java
@@ -1,9 +1,10 @@
 package baseball.version1;
 
-import static baseball.Constants.*;
+import static baseball.version1.Constants.DIGIT_NUMBER;
+import static baseball.version1.Constants.RANGE_END_NUMBER;
+import static baseball.version1.Constants.RANGE_START_NUMBER;
 
 import camp.nextstep.edu.missionutils.Randoms;
-
 import java.util.ArrayList;
 
 public class Computer {
