+++ src/main/java/baseball/version1/Constants.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.version1;
 
 public class Constants {
