+++ src/main/java/baseball/version1/DataValidator.java
@@ -1,6 +1,8 @@
 package baseball.version1;
 
-import static baseball.Constants.*;
+import static baseball.version1.Constants.CONTINUE_NUMBER;
+import static baseball.version1.Constants.DIGIT_NUMBER;
+import static baseball.version1.Constants.STOP_NUMBER;
 
 import java.util.ArrayList;
 import java.util.HashSet;
