+++ src/main/java/baseball/version2/Constants/Phrases.java
@@ -2,7 +2,7 @@ package baseball.version2.Constants;
 
 public class Phrases {
 
-    public final static String START_PHRASES = "숫자 야구 게임을 시작합니다";
+    public final static String START_PHRASES = "숫자 야구 게임을 시작합니다.";
 
     public final static String INSERT_ANSWER_PHRASES = "숫자를 입력해주세요 : ";
 
@@ -15,6 +15,7 @@ public class Phrases {
     public final static String BALL_PHRASES = "볼";
     public final static String STRIKE_PHRASES = "스트라이크";
 
-    public final static String ERROR_PHRASES = "1~9까지의 서로 다른 세 자리 숫자만 입력하세요";
+    public final static String ANSWER_ERROR_PHRASES = "1~9까지의 서로 다른 세 자리 숫자만 입력하세요";
+    public final static String DECISION_ERROR_PHRASES = "1 또는 2를 입력해주세요";
 
 }
